DiscordBot/requirements.txt
discord.py
mysql-connector-python
python-dotenv
bcrypt
mariadb